deliveries_sds/requirements.txt
django-cors-headers==4.1.0
tzlocal
pytz